Validation demonstration
This commit shows how validation can be added very easily. It defines
a custom JSR-303 constraint (see `@Version`) that is added to the
`@ModelAttributes` class used by the action. If you put rubbish value,
you can see a validation error now.

If you put a more recent version than the original version, nothing was
happening really. This commit updates the loader to use
`VersionStringComparator` that throws a specific exception. Said
exception is annotated so that it returns a `400` (bad request) rather
than a "generic" 500.

Finally, this commit registers a custom error page so that we show a
"meaningful" error page rather that the vanilla error page.