Devtools demonstration
To enable DevTools, simply add the dependency to your project. If you
restart the app, a LiveReload server is automatically started and you
can connect your browser to the application.

Previously, if you diffed `1.2.5.RELEASE` and `1.3.0.M3`, nothing was
indicating that the `spring.view.*` keys are deprecated. This commit
fixes that by adding a CSS entry:

```
.deprecated {
    text-decoration: line-through
}
```

No restart is required. This commit also filters out the groups that have
not changed. For that we need to change the template to offer an
additional option (a checkbox). And we need to handle the new boolean
value on the server-side, filtering the collection if necessary. All this
was taken into account live (no need to restart the server).

Finally, if we go back on the home page, we had an empty form. It would
be nice if it was pre-filled with some data. This commit updates the
controller to set some values in there. It should show up immediately in
the browser since LiveReload makes sure to refresh the browser once the
compilation has completed.

To sum up, this commit updates CSS, HTML and Java code without (cold)
restarting the server.