Cache demonstration
While our library has obviously some sort of caching of the Maven
meta-data we can see in the logs that every time we need to diff two
versions, it is called to recompute the model.

We can add the cache starter and `@EnableCaching` on our app and it will
just work out-of-the-box. Since we have no caching provider available,
Spring Boot uses a simple implementation based on concurrent hash maps.

We can also show how we can integrate a real cache library. For the sake
of the demo, we're going to use ehcache 3 and create a
JCacheManagerCustomizer that creates our `diffs` cache. Since the model
may refer to a SNAPSHOT version, we need to expire the entry after some
time. So we create a custom configuration that does that.

As we have the actuator now, we can see that the `/metrics` endpoint now
shows hit/miss stats of our cache.