Switch to AngularJS application
This commit transforms our mustache-based web application to AngularJS
using thymeleaf. This adds a `client` module that uses NPM to gather the
necessary client-side libraries.

It represents a different way of developping web applications where the
client code is completely separated from the server code. The purpose
here is to show that Spring Boot will work equally as well in such
arrangement.

The `client` module is now a webjar as any other client asset you may
want to use. The server now depends on that dependency so we need to
make sure to "install" it first.