Caching and compression demonstration
By default, the server does not compress any response. We can easily
switch that on via the `spring.compression` namespace and only compress
JSON, CSS and Javascript.

We can see via Chrome Devtools that the json response is significantly
smaller now. But we could refrain from sending unecessary data. One way
to achieve that is by creating a json view for our endpoint. We can
annotate our model with only the fields that the client actually needs.

This is better but we could also cache the whole document in the browser
itself. For that we can create an etag that defines the from and to
versions. Because our library is resolving the actual version when it
needs to provide the diff for a SNAPSHOT, the etag is truely unique.