Auto-completion demonstration
This commit showcases an end-to-end development of a feature using
devtools (no restart necessary). It updates the client to offer a list
of available releases to pick from (auto-completion).

We need to create an extra endpoint that's going to provide the values
and update the client to contact that new service to initialize the
auto-completion list.

To prevent calls on a remote service, the result can be cached as well.