Static resource demonstration
This commit showcases how the same friendly developer experience is
possible even if some resources are located in a different module (CSS
and Javascript).

We need to update our configuration so that Spring Boot looks in those
additional folders: that way, we can edit the source and have immediate
feedback in the application. To make that consistent, we will specify
folders that are relative to the working directory of the application,
that is the `spring-boot-4tw-web` module. We can make sure our local run
configuration uses that by adding `$MODULE_DIR` as working directory in
IntelliJ IDEA. Build tools are using that directory by default.

We're doing that in a `dev` profile so that these extra local locations
aren't added to the app by default. We also teach devtools to trigger a
LiveReload refresh when such files change.

The new application processes the diff in the background and has no
visual indication that it is busy. This commit changes the client code
to add a visual effect.