Optimize static assets demonstration
In production we may want to optimize client resources. We can use
gulp to minify javascript and CSS resources that are bundled in the
webjar. We enable resource caching in a `production` profile.

In `dev` mode, nothing changes: the client still loads the code from the
`src` and `bower_components` directories.